added estimation history label, fixed percentages in donut chart, fixed update estimation in add owner by using ownertoupdate.munID, tweaked some entities.

+++ lib/HOMEPAGES2/entities2/new/Municipality.dart
@@ -10,7 +10,7 @@ class Municipality {
   double totalDamageCost;
   List<Owner> owners = [];
   Color colorMarker;
-  int? damageCostInPercentage;
+  double? damageCostInPercentage;
 
   double? highestWs;
   double? highestRf24;
